requirements.txt
PyYAML~=6.0.1
Flask~=3.0.2
gunicorn~=21.2.0
flask_compress
gitpython